pyproject.toml
[project]
name = "python-drf-book-shop"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.9",
]
